+++ mock/fn.ts
@@ -2,18 +2,114 @@
 // This module is browser compatible.
 
 import { Mock } from "./mock.ts";
+import type { MockSpec } from "./mock.ts";
 import { incrementalNumber } from "./utils.ts";
-import type { MockObject } from "./mock.ts";
 
+interface MockObject<A extends readonly unknown[] = any[], R = unknown> {
+  (...args: A): R;
+  mock: Pick<MockSpec, "calls" | "results" | "callOrderNumbers">;
+
+  /** Sets the mock function as default. The set function will be called when the mock
+   * object is called.
+   * ```ts
+   * import { expect, fn, test } from "https://deno.land/x/unitest@$VERSION/mod.ts";
+   *
+   * test("should define implementation as default", () => {
+   *   const mockObject = fn().defaultImplementation(() => true);
+   *   expect(mockObject()).toBe(true);
+   * });
+   * ```
+   */
+  defaultImplementation(
+    implementation: (...args: A) => R,
+  ): MockObject<A, R>;
+
+  /** Sets default as return value. The set value will be return when the mock object
+   * is called.
+   * ```ts
+   * import { expect, fn, test } from "https://deno.land/x/unitest@$VERSION/mod.ts";
+   *
+   * test("should define return value as default", () => {
+   *   const mockObject = fn(() => 1).defaultReturnValue(0);
+   *   expect(mockObject()).toBe(0);
+   * });
+   * ```
+   */
+  defaultReturnValue(value: R): MockObject<A, R>;
+
+  /** Sets default as resolved value. The set value will be Promised and return when
+   * the mock object is called.
+   * ```ts
+   * import { expect, fn, test } from "https://deno.land/x/unitest@$VERSION/mod.ts";
+   *
+   * test("should define return value as default", () => {
+   *   const mockObject = fn().defaultResolvedValue(1);
+   *   expect(mockObject()).toEqual(Promise.resolve(1));
+   * });
+   * ```
+   */
+  defaultResolvedValue(value: R): MockObject<A, Promise<R>>;
+
+  /** Sets a mock function to be called only once. This takes precedence over the
+   * default mock function. If there is more than one once implementation, they will
+   * be called in the order of registration.
+   * ```ts
+   * import { expect, fn, test } from "https://deno.land/x/unitest@$VERSION/mod.ts";
+   *
+   * test("should define implementation as only once", () => {
+   *   const mockObject = fn(() => 0).onceImplementation(() => 1);
+   *   expect(mockObject()).toBe(1);
+   *   expect(mockObject()).toBe(0);
+   * });
+   * ```
+   */
+  onceImplementation(
+    implementation: (...args: A) => R,
+  ): MockObject<A, R>;
+
+  /** Sets a mock function what return specific value to be called only once. This
+   * takes precedence over the default mock function. Follow the FIFO.
+   * ```ts
+   * import { expect, fn, test } from "https://deno.land/x/unitest@$VERSION/mod.ts";
+   *
+   * test("should define return value as only once", () => {
+   *   const mockObject = fn(() => 1).onceReturnValue(0);
+   *   expect(mockObject()).toBe(0);
+   *   expect(mockObject()).toBe(1);
+   * });
+   * ```
+   */
+  onceReturnValue(value: R): MockObject<A, R>;
+}
+
+/** store fn internal implementation */
+class MockFnStore {
+  private onceImplementations: ((...args: unknown[]) => unknown)[] = [];
+  constructor(
+    private defaultImplementation?: ((...args: unknown[]) => unknown),
+  ) {}
+
+  /** pick implementation as FIFO */
+  pickImplementation(): ((...args: unknown[]) => unknown) | undefined {
+    return this.onceImplementations.shift() ?? this.defaultImplementation;
+  }
+}
+
+/** make mock object with implementation function */
+function fn<A extends readonly unknown[], R>(
+  implementation: (...args: A) => R,
+): MockObject<A, R>;
 /** make mock object */
-function fn<T extends unknown[]>(
-  implementation: (...args: T) => unknown,
-): MockObject<T>;
 function fn(): MockObject;
-function fn(implementation?: (...args: unknown[]) => unknown): MockObject {
+function fn(
+  implementation?: (...args: unknown[]) => unknown,
+): MockObject {
   const mock = new Mock();
+  const mockFnStore = new MockFnStore(implementation);
 
-  const call = (...args: unknown[]): void => {
+  /** Calls a method of an object, substituting another object for the current object.  */
+  const call = (...args: unknown[]): unknown => {
+    const implementation = mockFnStore.pickImplementation();
     const value = implementation?.(...args);
     mock.add({
       args,
@@ -23,6 +119,50 @@ function fn(implementation?: (...args: unknown[]) => unknown): MockObject {
       },
       orderNumber: incrementalNumber(),
     });
+    return value;
+  };
+
+  /** Sets the mock function as default. The set function will be called when the mock object is called.  */
+  const defaultImplementation = (
+    implementation: (...args: unknown[]) => unknown,
+  ): MockObject => {
+    mockFnStore["defaultImplementation"] = implementation;
+    return call as MockObject;
+  };
+
+  /** Sets a mock function to be called only once.
+   * This takes precedence over the default mock function.
+   * If there is more than one once implementation, they will be called in the order of registration.
+   */
+  const onceImplementation = (
+    implementation: (...args: unknown[]) => unknown,
+  ): MockObject => {
+    mockFnStore["onceImplementations"].push(implementation);
+    return call as MockObject;
+  };
+
+  /** Sets a mock function what return specific value to be called only once.
+   * This takes precedence over the default mock function.
+   * Follow the FIFO.
+   */
+  const onceReturnValue = (value: unknown): MockObject => {
+    mockFnStore["onceImplementations"].push(() => value);
+    return call as MockObject;
+  };
+
+  /** Sets default as return value. The set value will be return when the mock object is called.
+   */
+  const defaultReturnValue = (value: unknown): MockObject => {
+    mockFnStore["defaultImplementation"] = () => value;
+    return call as MockObject;
+  };
+
+  /** Sets default as resolved value.
+   * The set value will be Promised and return when the mock object is called.
+   */
+  const defaultResolvedValue = (value: unknown): MockObject => {
+    mockFnStore["defaultImplementation"] = () => Promise.resolve(value);
+    return call as MockObject;
   };
 
   Object.defineProperty(call, "mock", {
@@ -32,8 +172,26 @@ function fn(implementation?: (...args: unknown[]) => unknown): MockObject {
     },
   });
 
+  const properties = Object.entries({
+    defaultImplementation,
+    defaultReturnValue,
+    defaultResolvedValue,
+    onceImplementation,
+    onceReturnValue,
+  }).reduce(
+    (acc, [key, value]) => ({
+      ...acc,
+      [key]: {
+        value,
+      },
+    }),
+    {} as PropertyDescriptorMap,
+  );
+
+  Object.defineProperties(call, properties);
+
   return call as MockObject;
 }
 
-export { fn };
+export { fn, MockFnStore };
 export type { MockObject };